SELECT Dynamic Resizing Tests
Removed the special CSS from the tests, because it was affecting negatively (in unexpected ways), the way the SELECT javascript-drive height resizing/adjustment works.
I need to later understand how some of those CSS properties ("overflow", "min-height" on the SELECT element) affect JS in isolation of each other.

Added a JS test for whether the SELECT Fancy Widget is enabled (e.g. for most mobile device browsers).  The assumption is that if then default OPTION elements are not available for selection/clicking (i.e. Fancy Widget is handling the selection), their CSS (inner) height (i.e. height - margin/border) will be 0.  I create a SELECT element in a controlled manner on the fly and measure its OPTION elements heights. Because it's a controlled manner, I can be sure to make the margin/border/padding to be zero (0).
NOTE: The modern cool Internet "hot" tools (e.g. Modernizr) don't offer such tests. So I had to make my own.  It is based on assumptions, but it works for my special case/scenario.